requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-generativeai==0.3.2
requests==2.31.0
gunicorn==22.0.0
python-docx==1.1.2
python-dotenv==1.0.1
